[Impeller] Place Entity Geometry classes in separate TUs
