+++ impeller/aiks/canvas.h
@@ -16,7 +16,8 @@
 #include "impeller/aiks/picture.h"
 #include "impeller/core/sampler_descriptor.h"
 #include "impeller/entity/entity_pass.h"
-#include "impeller/entity/geometry.h"
+#include "impeller/entity/geometry/geometry.h"
+#include "impeller/entity/geometry/vertices_geometry.h"
 #include "impeller/geometry/matrix.h"
 #include "impeller/geometry/path.h"
 #include "impeller/geometry/point.h"
