+++ impeller/entity/BUILD.gn
@@ -223,8 +223,20 @@ impeller_component("entity") {
     "entity_pass_delegate.h",
     "entity_pass_target.cc",
     "entity_pass_target.h",
-    "geometry.cc",
-    "geometry.h",
+    "geometry/cover_geometry.cc",
+    "geometry/cover_geometry.h",
+    "geometry/fill_path_geometry.cc",
+    "geometry/fill_path_geometry.h",
+    "geometry/geometry.cc",
+    "geometry/geometry.h",
+    "geometry/point_field_geometry.cc",
+    "geometry/point_field_geometry.h",
+    "geometry/rect_geometry.cc",
+    "geometry/rect_geometry.h",
+    "geometry/stroke_path_geometry.cc",
+    "geometry/stroke_path_geometry.h",
+    "geometry/vertices_geometry.cc",
+    "geometry/vertices_geometry.h",
     "inline_pass_context.cc",
     "inline_pass_context.h",
   ]
