+++ impeller/entity/geometry/fill_path_geometry.h
@@ -1,0 +1,42 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#pragma once
+
+#include "impeller/entity/geometry/geometry.h"
+
+namespace impeller {
+
+/// @brief A geometry that is created from a filled path object.
+class FillPathGeometry : public Geometry {
+ public:
+  explicit FillPathGeometry(const Path& path);
+
+  ~FillPathGeometry();
+
+ private:
+  // |Geometry|
+  GeometryResult GetPositionBuffer(const ContentContext& renderer,
+                                   const Entity& entity,
+                                   RenderPass& pass) override;
+
+  // |Geometry|
+  GeometryVertexType GetVertexType() const override;
+
+  // |Geometry|
+  std::optional<Rect> GetCoverage(const Matrix& transform) const override;
+
+  // |Geometry|
+  GeometryResult GetPositionUVBuffer(Rect texture_coverage,
+                                     Matrix effect_transform,
+                                     const ContentContext& renderer,
+                                     const Entity& entity,
+                                     RenderPass& pass) override;
+
+  Path path_;
+
+  FML_DISALLOW_COPY_AND_ASSIGN(FillPathGeometry);
+};
+
+}  // namespace impeller
