+++ impeller/entity/geometry/vertices_geometry.cc
@@ -1,0 +1,11 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#include "impeller/entity/geometry/vertices_geometry.h"
+
+namespace impeller {
+
+//
+
+}  // namespace impeller
